Move plugin android code into sub packages to avoid clashes
