+++ plugins/flutter_aepassurance/android/src/main/java/com/adobe/marketing/mobile/flutter/flutter_aepassurance/FlutterAEPAssurancePlugin.java
@@ -9,7 +9,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.flutter;
+package com.adobe.marketing.mobile.flutter.flutter_aepassurance;
 
 import android.util.Log;
